+++ src/main/java/com/raoulvdberge/refinedstorage/integration/inventorysorter/IntegrationInventorySorter.java
@@ -6,5 +6,6 @@ public class IntegrationInventorySorter {
     public static void register() {
         // Prevent items moving while scrolling through slots with Inventory Sorter in the Crafter Manager
         FMLInterModComms.sendMessage("inventorysorter", "containerblacklist", "com.raoulvdberge.refinedstorage.container.ContainerCrafterManager");
+        FMLInterModComms.sendMessage("inventorysorter", "slotblacklist", "com.raoulvdberge.refinedstorage.container.slot.SlotCrafterManager");
     }
 }